backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==4.9.0
qdrant-client==1.8.0
openai==1.3.5
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
trafilatura==1.6.2